mobile version concluded
